+++ src/io/trr_io.rs
@@ -97,7 +97,7 @@ impl FrameData for TrrFrameData {
     /// Update the `System` structure based on data from `TrrFrameData`.
     fn update_system(self, system: &mut System) {
         unsafe {
-            for (i, atom) in system.get_atoms_as_ref_mut().iter_mut().enumerate() {
+            for (i, atom) in system.get_atoms_as_mut().iter_mut().enumerate() {
                 let pos = Vector3D::from(*self.coordinates.get_unchecked(i));
                 let vel = Vector3D::from(*self.velocities.get_unchecked(i));
                 let force = Vector3D::from(*self.forces.get_unchecked(i));
@@ -1967,10 +1967,7 @@ mod tests {
 
         // remove the protein group from the system; this should not change the output of the XtcGroupWriter
         unsafe {
-            let val = system
-                .get_groups_as_ref_mut()
-                .swap_remove("Protein")
-                .unwrap();
+            let val = system.get_groups_as_mut().swap_remove("Protein").unwrap();
             assert_eq!(val.get_atoms(), writer.group.get_atoms());
             assert!(!system.group_exists("Protein"));
         }